update names for scripts
